requirements.txt
SQLAlchemy~=1.4.2
Werkzeug~=1.0.1
WTForms~=2.3.3
Flask~=1.1.2
Flask-RESTful==0.3.8
python-dotenv~=0.14.0
Flask-Login==0.5.0
Flask-WTF==0.14.3
SQLAlchemy-serializer==1.3.4.4
requests==2.25.0
email-validator==1.1.2
python-dotenv==0.14.0